Move st_exception to exception_test_file
